Added computed filter function for Routine category
Also stripped out very muddied logic from view template
for the template being rendered for unfiltered and
filtered routine list.

The next step would be to filter by duration and frequencies
while showing the other two properties in the list.

The addition of the computed filter function helped the
view to be data-driven. And, there was one less `if` condition
appearing in the view.