The ID app was repeated twice, removed from App.vue
An ID should appear only once per page. It so happened
due to oversight, that the ID app was appearing once
in the index.html and also in the App.vue template as
a child of the div.row element.

This was causing Vue's Root element to not be set and
that could later have caused other inconsistencies.
(Perhaps, Vue should warn about repeating ID elements,
but that's a tall order...)

We have chosen to remove the ID from the App.vue template,
this means that the root element is div#app and Bootstrap's
div.container is its immediate child. I don't expect that
there will be sibling to the div.container tag.

The entire app is thus rooted at div#app defined in index.html,
App.vue defines the outer shell of the Vue app that then is
managed via the Router - with the <router-view/> being the
child of the .row whose parent is .container under #app.

This solves a major skew that occurred due to lack of Vue
root element definition, due to repeated #app definitions.